Refactor zosmfjobservice into class runners
