--- jobs-api-server/src/main/java/org/zowe/jobs/services/ZoweApiRestExceptionReturner.java
@@ -1,19 +1,0 @@
-/*
- * This program and the accompanying materials are made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Copyright IBM Corporation 2019
- */
-package org.zowe.jobs.services;
-
-import com.google.gson.JsonObject;
-
-import org.zowe.api.common.exceptions.ZoweApiRestException;
-
-@FunctionalInterface
-public interface ZoweApiRestExceptionReturner {
-    ZoweApiRestException run(JsonObject jsonResponse);
-}
